Move asset server and gateway server to isolated package
