+++ run-gateway.sh
@@ -2,5 +2,5 @@
 
 pushd $(dirname "${0}") > /dev/null
 cd run
-yarn node ../services/main/gateway-server.mjs --service-name=$1
+yarn node ../services/gateway/gateway-server.mjs --service-name=$1
 popd > /dev/null
